+++ src/commands/Refresh.js
@@ -19,7 +19,7 @@ class Refresh extends Command {
     }
 
     debug('Loaded config:')
-    debug(this.runtime.compile)
+    debug(this.runtime.config)
 
     terraformArgs.push(`-state=${this.runtime.init.paths.stateFile}`)