+++ test/scenario/compile/run.sh
@@ -36,7 +36,7 @@ PATH=${PATH} \
 USER=${USER} \
 HOME=${HOME} \
 FREY_SHOULD_BE_AS_VAR_IN_TERRAFORM=now \
-"${__node}" "${__root}/${__codelib}/cli.js" compile \
+"${__node}" "${__root}/${__codelib}/cli.js" config \
   --cfg-var "global.ssh.keysdir=${__sysTmpDir}" \
   --cfg-var "infra.provider.aws.region=eu-west-1" \
   --verbose \